Updated AspspServiceImpl with if-statements to check whether the path to custom file has been provided, custom file exists and is writable.

+++ xs2a-adapter-aspsp-registry/src/main/java/de/adorsys/xs2a/adapter/registry/AspspServiceImpl.java
@@ -3,9 +3,14 @@ package de.adorsys.xs2a.adapter.registry;
 import de.adorsys.xs2a.adapter.service.AspspCsvService;
 import de.adorsys.xs2a.adapter.service.AspspRepository;
 import de.adorsys.xs2a.adapter.service.AspspService;
+import de.adorsys.xs2a.adapter.service.PropertyUtil;
 import de.adorsys.xs2a.adapter.service.model.Aspsp;
 
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.List;
+
+import static de.adorsys.xs2a.adapter.registry.AspspCsvServiceImpl.CSV_ASPSP_ADAPTER_CONFIG_FILE_PATH;
 
 public class AspspServiceImpl implements AspspService {
     private final AspspRepository aspspRepository;
@@ -38,6 +43,22 @@ public class AspspServiceImpl implements AspspService {
 
     @Override
     public void importAspsps(List<Aspsp> aspsps) {
+        String pathToCsvFile = PropertyUtil.readProperty(CSV_ASPSP_ADAPTER_CONFIG_FILE_PATH);
+
+        if (pathToCsvFile.isEmpty()) {
+            throw new RuntimeException(
+                String.format(
+                    "Import operation could be proceeded on custom CSV file only. " +
+                        "Path to custom file should be provided through [%s] property.",
+                    CSV_ASPSP_ADAPTER_CONFIG_FILE_PATH
+                )
+            );
+        }
+
+        if (!Files.isWritable(Paths.get(pathToCsvFile))) {
+            throw new RuntimeException(String.format("File [%s] does not exist or is not writable", pathToCsvFile));
+        }
+
         aspspRepository.deleteAll();
         aspspRepository.saveAll(aspsps);
         aspspCsvService.saveCsv();
